FIX: Bad timing of UI state attaching.
UI state was being attached to model by observing model event notifications.
That was fragile because some other component also can observe them and try to access some UI state at the point. Because multicasting order is undefined, it can try to access for UI state that is not yet attached.

This patch fixes this by using Objective-C associated object instead of event observation. UI state will be attached at first access to it, and destroyed when the host model node object dies.